requirements.txt
transformers
transformers[sentencepiece]
torch
pandas
scikit-learn
nltk
pinecone-client
seaborn
matplotlib
tqdm
bs4
py7zr
tqdm
PyYAML
notebook
boto3
streamlit
requests
PyPDF2
Pillow
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
bentoml==1.3.9
-e .